fix product list
